requirements.txt
langchain_core
langchain_groq
langgraph
langgraph_supervisor
langchain_ollama
dotenv
psycopg2-binary
langchain
# langgraph_checkpoints
ipython
typing_extensions
streamlit

langchain
langgraph
langchain-community

psycopg2-binary
pymysql
pymongo
redis

groq
ollama

pandas
plotly
numpy
scikit-learn
scipy

PyJWT
python-dotenv
bcrypt
cryptography

requests
httpx

celery
rq

pytest
pytest-cov
pytest-mock

black
flake8
pre-commit

prometheus-client
opentelemetry-api
opentelemetry-sdk